add type

+++ entity/referral.go
@@ -3,8 +3,8 @@ package entity
 import "time"
 
 type ReferralDb struct {
-	ID         uint `gorm:"primaryKey"`
-	UserID     uint `gorm:"unique"`
-	Caller     uint
+	ID         uint   `gorm:"primaryKey"`
+	UserID     uint64 `gorm:"unique"`
+	Caller     uint64
 	TimeCreate time.Time
 }
